+++ lib/src/models/project_model.dart
@@ -2,6 +2,8 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:fvm/constants.dart';
+import 'package:fvm/src/services/config_repository.dart';
+import 'package:fvm/src/services/pubspec_repository.dart';
 import 'package:fvm/src/utils/extensions.dart';
 import 'package:path/path.dart';
 import 'package:pubspec2/pubspec2.dart';
@@ -11,27 +13,27 @@ import 'config_model.dart';
 /// Flutter Project model
 class Project {
   /// Directory of project
-  final Directory projectDir;
+  final String path;
 
   /// Config found within the project
-  ProjectConfig? config;
+  ConfigDto? config;
 
   final PubSpec? pubspec;
 
   /// Project constructor
   Project({
     required this.config,
-    required this.projectDir,
+    required this.path,
     required this.pubspec,
   });
 
   /// Returns the project name
-  String get name => basename(projectDir.path);
+  String get name => basename(path);
 
   /// Pinned version within a project
   /// returns null if no version is pinned
   String? get pinnedVersion {
-    return config?.flutter;
+    return config?.flutterSdkVersion;
   }
 
   /// Returns the active configured flavor
@@ -59,18 +61,15 @@ class Project {
     return pubspec?.dependencies.containsKey('flutter') ?? false;
   }
 
-  Directory get fvmDir {
-    return Directory(join(projectDir.path, kFvmDirName));
-  }
+  Directory get fvmPath => Directory(_getLocalFvmPath(path));
 
-  Directory get fvmCacheDir {
-    return Directory(join(fvmDir.path, 'versions'));
-  }
+  Directory get fvmCachePath =>
+      Directory(join(_getLocalFvmPath(path), 'versions'));
 
   /// Returns the project path to the Flutter SDK symlink
   Link get cacheVersionSymlink {
     return Link(join(
-      fvmCacheDir.path,
+      fvmCachePath.path,
       pinnedVersion,
     ));
   }
@@ -78,20 +77,18 @@ class Project {
   /// Compatibility version for non-admin permission
   Directory get cacheVersionSymlinkCompat {
     return Directory(join(
-      fvmCacheDir.path,
+      fvmCachePath.path,
       pinnedVersion,
     ));
   }
 
   /// .gitignore file
-  File get gitignoreFile {
-    return File(join(projectDir.path, '.gitignore'));
-  }
+  File get gitignoreFile => File(join(path, '.gitignore'));
 
   /// Old linking path
   Link get legacyCacheVersionSymlink {
     return Link(join(
-      projectDir.path,
+      path,
       kFvmDirName,
       'flutter_sdk',
     ));
@@ -99,23 +96,42 @@ class Project {
 
   /// Returns dart tool package config
   File get _dartToolPackgeConfig {
-    return File(join(projectDir.path, '.dart_tool', 'package_config.json'));
+    return File(join(path, '.dart_tool', 'package_config.json'));
   }
 
   File get _dartToolVersionFile {
-    return File(join(projectDir.path, '.dart_tool', 'version'));
+    return File(join(path, '.dart_tool', 'version'));
   }
 
   /// Pubspec file
-  File get pubspecFile {
-    return File(join(projectDir.path, 'pubspec.yaml'));
-  }
+  String get pubspecPath => join(path, 'pubspec.yaml');
 
   /// Config file
-  File get configFile {
-    return File(join(fvmDir.path, kFvmConfigFileName));
-  }
+  String get configPath => _getLocalFvmConfigPath(path);
 
   /// Checks if project has config
   bool get hasConfig => config != null;
+
+  static Project loadFromPath(String path) {
+    final configPath = _getLocalFvmConfigPath(path);
+
+    final config = ConfigRepository(configPath).load();
+
+    final pubspecPath = join(path, 'pubspec.yaml');
+    final pubspec = PubspecRepository(pubspecPath).load();
+
+    return Project(
+      path: path,
+      pubspec: pubspec,
+      config: config,
+    );
+  }
+}
+
+String _getLocalFvmPath(String path) {
+  return join(path, kFvmDirName);
+}
+
+String _getLocalFvmConfigPath(String path) {
+  return join(_getLocalFvmPath(path), kFvmConfigFileName);
 }