+++ lib/src/services/project_service.dart
@@ -9,7 +9,6 @@ import 'package:fvm/src/utils/logger.dart';
 import 'package:fvm/src/utils/pretty_json.dart';
 import 'package:mason_logger/mason_logger.dart';
 import 'package:path/path.dart';
-import 'package:pubspec2/pubspec2.dart';
 
 import '../../constants.dart';
 
@@ -20,37 +19,25 @@ class ProjectService {
 
   static ProjectService get instance => ctx.get<ProjectService>();
 
-  /// Returns projects by providing a [directory]
-  Future<Project> loadByDirectory(Directory directory) async {
-    final configFile = File(
-      join(directory.path, kFvmDirName, kFvmConfigFileName),
-    );
+  /// Recursive look up to find nested project directory
+  /// Can start at a specific [directory] if provided
+  Future<Project> findAncestor({Directory? directory}) async {
+    // Get directory, defined root or current
+    directory ??= Directory(ctx.workingDirectory);
 
-    final pubspecFile = File(join(directory.path, 'pubspec.yaml'));
+    // Checks if the directory is root
+    final isRootDir = rootPrefix(directory.path) == directory.path;
 
-    ProjectConfig? config;
+    // Gets project from directory
+    final project = Project.loadFromPath(directory.path);
 
-    if (configFile.existsSync()) {
-      final jsonString = configFile.readAsStringSync();
-      config = ProjectConfig.fromJson(jsonString);
-    }
+    // If project has a config return it
+    if (project.hasConfig) return project;
 
-    PubSpec? pubspec;
-    if (pubspecFile.existsSync()) {
-      final jsonString = pubspecFile.readAsStringSync();
-      pubspec = PubSpec.fromYamlString(jsonString);
-    }
+    // Return working directory if has reached root
+    if (isRootDir) return Project.loadFromPath(ctx.workingDirectory);
 
-    return Project(
-      projectDir: directory,
-      pubspec: pubspec,
-      config: config,
-    );
-  }
-
-  /// Returns a list of projects by providing a list of [paths]
-  Future<List<Project>> fetchProjects(List<Directory> paths) async {
-    return Future.wait(paths.map(loadByDirectory));
+    return await findAncestor(directory: directory.parent);
   }
 
   /// Adds to .gitignore paths that should be ignored for fvm
@@ -115,10 +102,10 @@ class ProjectService {
       project.legacyCacheVersionSymlink.deleteSync();
     }
 
-    if (project.fvmCacheDir.existsSync()) {
-      project.fvmCacheDir.deleteSync(recursive: true);
-      project.fvmCacheDir.createSync(recursive: true);
+    if (project.fvmCachePath.existsSync()) {
+      project.fvmCachePath.deleteSync(recursive: true);
     }
+    project.fvmCachePath.createSync(recursive: true);
 
     createLink(
       project.cacheVersionSymlink,
@@ -136,7 +123,7 @@ class ProjectService {
     Project project,
   ) {
     final vscodeDir = Directory(join(
-      project.projectDir.path,
+      project.path,
       '.vscode',
     ));
 
@@ -202,32 +189,11 @@ class ProjectService {
 
     final relativePath = relative(
       project.cacheVersionSymlinkCompat.path,
-      from: project.projectDir.path,
+      from: project.path,
     );
 
     currentSettings["dart.flutterSdkPath"] = relativePath;
 
     vsCodeSettingsFile.writeAsStringSync(prettyJson(currentSettings));
-  }
-
-  /// Recursive look up to find nested project directory
-  /// Can start at a specific [directory] if provided
-  Future<Project> findAncestor({Directory? directory}) async {
-    // Get directory, defined root or current
-    directory ??= Directory(ctx.workingDirectory);
-
-    // Checks if the directory is root
-    final isRootDir = rootPrefix(directory.path) == directory.path;
-
-    // Gets project from directory
-    final project = await loadByDirectory(directory);
-
-    // If project has a config return it
-    if (project.hasConfig) return project;
-
-    // Return working directory if has reached root
-    if (isRootDir) return loadByDirectory(Directory(ctx.workingDirectory));
-
-    return await findAncestor(directory: directory.parent);
   }
 }