+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Flutter repo git reference for much faster clone of new versions.
 - Removed --skip-setup on `use` command. If you need to install without setting up, you should use `install`. You can force SDK setup on install by using -s or --setup
 
+### Breaking changes
+
+- FVM Config - FVM config now will be saved as fvm.yml at the root of the project. This will allow for better project configuration and versioning, `.fvm/fvm_config.json` will still be generated to be used by FVM, but it can be added to `.gitignore`. It will be generated when runing a `fvm install` or `fvm use` command.
+
 ## 2.4.1
 
 - Filter out Mac releases based on architecture.